Seperate Trading and Backtester Interface
