+++ src/Mynt.Core/Backtester/BackTestRunner.cs
@@ -10,7 +10,7 @@ namespace Mynt.Core.Backtester
 {
     public class BackTestRunner
     {
-        public async Task<List<BackTestResult>> RunSingleStrategy(ITradingStrategy strategy, BacktestOptions backtestOptions, IDataStore dataStore)
+        public async Task<List<BackTestResult>> RunSingleStrategy(ITradingStrategy strategy, BacktestOptions backtestOptions, IDataStoreBacktest dataStore)
         {
             var results = new List<BackTestResult>();
